+++ generators/app/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var yeoman = require('yeoman-generator');
+var Generator = require('yeoman-generator');
 
 var prompting = require('./src/prompting');
 var writing = require('./src/writing');
@@ -8,7 +8,19 @@ var configuring = require('./src/configuring');
 var install = require('./src/install');
 var end = require('./src/end');
 
-module.exports = yeoman.Base.extend({
+module.exports = Generator.extend({
+  constructor: function(args, opts) {
+    Generator.call(this, args, opts);
+
+    // Arguments
+    this.argument('pluginName', {
+      type: String,
+      desc: 'Optional plugin name to avoid being prompted',
+      required: false
+    });
+
+  },
+
   initializing: function() {
     this.properties = {
       templated: {},