+++ generators/app/src/configuring.js
@@ -2,14 +2,13 @@
 
 var chalk = require('chalk');
 var fs = require('fs');
-var paths = require('../config/paths');
 
 module.exports = function() {
-
   var self = this;
+  var done = this.async();
 
   // Define url for the OpenVeo plugin
-  this.properties.url = this.destinationPath(paths.openveo + this.properties.templated.plugin + '/');
+  this.properties.url = this.destinationPath('openveo-' + this.options.pluginName + '/');
 
   // Create plugin folder
   fs.mkdir(this.properties.url, function(err) {
@@ -17,7 +16,8 @@ module.exports = function() {
       self.env.error(chalk.red('The "' + self.properties.templated.Plugin + '" plugin already exist. \n'));
     } else if (err) {
       self.env.error(chalk.red(err));
-    }
+    } else
+      done();
   });
 
 };