+++ generators/app/templates/_Gruntfile.js
@@ -1,37 +1,28 @@
 'use strict';
 
-// Module dependencies
-var path = require('path');
-
-process.root<%= Plugin %> = __dirname;
-process.require<%= Plugin %> = function(filePath) {
-  return require(path.normalize(process.root<%= Plugin %> + '/' + filePath));
-};
+/* eslint no-sync: 0 */
+var fs = require('fs');
+var openVeoApi = require('@openveo/api');
+require('./processRequire.js');
 
 /**
  * Loads a bunch of grunt configuration files from the given directory.
  *
  * Loaded configurations can be referenced using the configuration file name.
- * For example, if myConf.js describes a property "test", it will be accessible
- * using myConf.test.
+ * For example, if myConf.js returns an object with a property "test", it will be accessible using myConf.test.
  *
- * @param String path Path of the directory containing configuration files
- * @return Object The list of configurations indexed by filename without
- * the extension
+ * @param {String} path Path of the directory containing configuration files
+ * @return {Object} The list of configurations indexed by filename without the extension
  */
 function loadConfig(path) {
-  var glob = require('glob');
-  var object = {};
-  var key;
+  var configuration = {};
+  var configurationFiles = fs.readdirSync(path);
 
-  glob.sync('*', {
-    cwd: path
-  }).forEach(function(option) {
-    key = option.replace(/\.js$/, '');
-    object[key] = require(path + '/' + option);
+  configurationFiles.forEach(function(configurationFile) {
+    configuration[configurationFile.replace(/\.js$/, '')] = require(path + '/' + configurationFile);
   });
 
-  return object;
+  return configuration;
 }
 
 module.exports = function(grunt) {
@@ -54,8 +45,9 @@ module.exports = function(grunt) {
   grunt.loadNpmTasks('grunt-eslint');
   grunt.loadNpmTasks('grunt-mkdocs');
   grunt.loadNpmTasks('grunt-gh-pages');
-  grunt.loadNpmTasks('grunt-rename');
-  grunt.loadNpmTasks('grunt-remove');
+
+  grunt.registerMultiTask('rename', openVeoApi.grunt.renameTask(grunt));
+  grunt.registerMultiTask('remove', openVeoApi.grunt.removeTask(grunt));
 
   // Listen to changes on SCSS files and generate CSS files
   grunt.registerTask('default', ['compass:dev', 'watch']);