+++ generators/app/templates/_conf.js
@@ -1,19 +1,31 @@
 'use strict';
 
 module.exports = {
-  routes: {
-    public: {
-    },
-    private: {
-    },
-    ws: {
+  http: {
+    routes: {
+      public: {
+      },
+      private: {
+      },
+      ws: {
+      }
+    }
+  },
+  socket: {
+    namespaces: {
+      public: {
+      },
+      private: {
+      }
     }
   },
   entities: {
   },
+  webServiceScopes: [
+  ],
   permissions: [
     {
-      id: '<%= plugin %>-access-todo-page',
+      id: '<%= originalPluginName %>-access-todo-page',
       name: '<%= PLUGIN %>.PERMISSIONS.ACCESS_TODO_PAGE_NAME'
     }
   ],
@@ -25,8 +37,8 @@ module.exports = {
         subMenu: [
           {
             label: '<%= PLUGIN %>.MENU.TODO',
-            path: '<%= plugin %>/todo',
-            permission: '<%= plugin %>-access-todo-page'
+            path: '<%= originalPluginName %>/todo',
+            permission: '<%= originalPluginName %>-access-todo-page'
           }
         ]
       }
@@ -39,16 +51,15 @@ module.exports = {
     scriptFiles: {
       base: [],
       dev: [
-        '/<%= plugin %>/ov<%= Plugin %>/<%= Plugin %>App.js',
-        '/<%= plugin %>/ov<%= Plugin %>/<%= Plugin %>Controller.js',
-        '/<%= plugin %>/ov<%= Plugin %>/<%= Plugin %>Service.js'
+        '/<%= originalPluginName %>/ov<%= Plugin %>/<%= Plugin %>App.js',
+        '/<%= originalPluginName %>/ov<%= Plugin %>/<%= Plugin %>Controller.js'
       ],
       prod: [
-        '/<%= plugin %>/be/js/openveo<%= Plugin %>.js'
+        '/<%= originalPluginName %>/be/js/openveo<%= Plugin %>.js'
       ]
     },
     cssFiles: [
-      '/<%= plugin %>/be/css/<%= plugin %>.css'
+      '/<%= originalPluginName %>/be/css/<%= plugin %>.css'
     ]
   }
 };