+++ generators/app/templates/tasks/mochaTest.js
@@ -9,7 +9,10 @@ module.exports = {
     options: {
       reporter: 'spec'
     },
-    src: ['tests/server/init.js', 'tests/server/*.js']
+    src: [
+      'tests/server/init.js',
+      'tests/server/plugin/*.js'
+    ]
   }
 
 };